Herramientas_Ultima_Generacion/airbnb-bigdata-project-main/src/functions/events/requirements.txt
flask==2.0.2
google-cloud-storage==1.43.0
google-cloud-error-reporting==1.4.1